refactor: Move remaining members of container-utils to telemetry-utils (#16972)
Requires adding a dependency from `telemetry-utils` on
`protocol-definitions`.

Moved API members:
* DataProcessingError 
* extractSafePropertiesFromMessage

The changes also attempt to better encapsulate / document parameter type
information for the 2 moved members by adding a type alias.

Also...
* All lingering references to and dependencies on container-utils have
been removed
* The package README has been updated to reflect its deprecated status
* Since the package no longer has tests, existing test infra has been
removed
