+++ layerInfo.json
@@ -61,7 +61,7 @@
 			},
 			"Telemetry-Utils": {
 				"packages": ["@fluidframework/telemetry-utils"],
-				"deps": ["Base-Utils", "Base-Definitions", "Core-Utils"]
+				"deps": ["Base-Utils", "Base-Definitions", "Core-Utils", "Protocol-Definitions"]
 			},
 			"Container-Utils": {
 				"packages": ["@fluidframework/container-utils"],
