+++ packages/dds/shared-object-base/src/sharedObject.ts
@@ -8,13 +8,13 @@ import { IFluidHandle, ITelemetryProperties } from "@fluidframework/core-interfa
 import {
 	ITelemetryLoggerExt,
 	createChildLogger,
+	DataProcessingError,
 	EventEmitterWithErrorHandling,
 	loggerToMonitoringContext,
 	MonitoringContext,
 	SampledTelemetryHelper,
 	tagCodeArtifacts,
 } from "@fluidframework/telemetry-utils";
-import { DataProcessingError } from "@fluidframework/container-utils";
 import { assert, EventEmitterEventType } from "@fluidframework/common-utils";
 import { AttachState } from "@fluidframework/container-definitions";
 import {
