+++ packages/loader/container-utils/package.json
@@ -30,35 +30,11 @@
 		"lint:fix": "npm run prettier:fix && npm run eslint:fix",
 		"prettier": "prettier --check . --ignore-path ../../../.prettierignore",
 		"prettier:fix": "prettier --write . --ignore-path ../../../.prettierignore",
-		"test": "npm run test:mocha",
-		"test:coverage": "nyc npm test -- --reporter xunit --reporter-option output=nyc/junit-report.xml",
-		"test:mocha": "mocha --ignore 'dist/test/types/*' --recursive dist/test -r node_modules/@fluidframework/mocha-test-setup --unhandled-rejections=strict",
-		"test:mocha:multireport": "npm run test:mocha",
-		"test:mocha:verbose": "cross-env FLUID_TEST_VERBOSE=1 npm run test:mocha",
+		"test": "echo \"This package is scheduled for deletion and now contains no tests.\" && exit 1",
 		"tsc": "tsc",
 		"tsc:watch": "tsc --watch",
 		"typetests:gen": "fluid-type-test-generator",
 		"typetests:prepare": "flub typetests --dir . --reset --previous --normalize"
-	},
-	"nyc": {
-		"all": true,
-		"cache-dir": "nyc/.cache",
-		"exclude": [
-			"src/test/**/*.ts",
-			"dist/test/**/*.js"
-		],
-		"exclude-after-remap": false,
-		"include": [
-			"src/**/*.ts",
-			"dist/**/*.js"
-		],
-		"report-dir": "nyc/report",
-		"reporter": [
-			"cobertura",
-			"html",
-			"text"
-		],
-		"temp-directory": "nyc/.nyc_output"
 	},
 	"dependencies": {
 		"@fluidframework/common-utils": "^1.1.1",
@@ -73,20 +49,11 @@
 		"@fluidframework/build-tools": "^0.22.0",
 		"@fluidframework/container-utils-previous": "npm:@fluidframework/container-utils@2.0.0-internal.6.1.1",
 		"@fluidframework/eslint-config-fluid": "^2.0.0",
-		"@fluidframework/mocha-test-setup": "workspace:~",
-		"@fluidframework/test-runtime-utils": "workspace:~",
 		"@microsoft/api-extractor": "^7.34.4",
-		"@types/mocha": "^9.1.1",
 		"@types/node": "^16.18.38",
 		"concurrently": "^7.6.0",
 		"copyfiles": "^2.4.1",
-		"cross-env": "^7.0.3",
 		"eslint": "~8.6.0",
-		"mocha": "^10.2.0",
-		"mocha-json-output-reporter": "^2.0.1",
-		"mocha-multi-reporters": "^1.5.1",
-		"moment": "^2.21.0",
-		"nyc": "^15.1.0",
 		"prettier": "~2.6.2",
 		"rimraf": "^4.4.0",
 		"typescript": "~4.5.5"
