+++ packages/loader/container-utils/src/test/tsconfig.json
@@ -3,7 +3,7 @@
 	"compilerOptions": {
 		"rootDir": "./",
 		"outDir": "../../dist/test",
-		"types": ["node", "mocha"],
+		"types": ["node"],
 		"declaration": false,
 		"declarationMap": false,
 		"skipLibCheck": true,
