+++ packages/runtime/container-runtime/src/containerRuntime.ts
@@ -33,6 +33,7 @@ import {
 	createChildLogger,
 	createChildMonitoringContext,
 	DataCorruptionError,
+	DataProcessingError,
 	GenericError,
 	raiseConnectedEvent,
 	PerformanceEvent,
@@ -42,7 +43,6 @@ import {
 	ITelemetryLoggerExt,
 	UsageError,
 } from "@fluidframework/telemetry-utils";
-import { DataProcessingError } from "@fluidframework/container-utils";
 import {
 	DriverHeader,
 	FetchSource,
