+++ packages/runtime/container-runtime/src/gc/garbageCollection.ts
@@ -5,7 +5,6 @@
 
 import { Timer } from "@fluidframework/common-utils";
 import { LazyPromise } from "@fluidframework/core-utils";
-import { DataProcessingError } from "@fluidframework/container-utils";
 import { IRequest, IRequestHeader } from "@fluidframework/core-interfaces";
 import {
 	gcTreeKey,
@@ -22,6 +21,7 @@ import {
 import {
 	createChildLogger,
 	createChildMonitoringContext,
+	DataProcessingError,
 	ITelemetryLoggerExt,
 	MonitoringContext,
 	PerformanceEvent,
