+++ packages/runtime/container-runtime/src/scheduleManager.ts
@@ -7,14 +7,12 @@ import { IDeltaManager } from "@fluidframework/container-definitions";
 import { IDocumentMessage, ISequencedDocumentMessage } from "@fluidframework/protocol-definitions";
 import {
 	createChildLogger,
-	ITelemetryLoggerExt,
 	DataCorruptionError,
-} from "@fluidframework/telemetry-utils";
-import { assert, performance } from "@fluidframework/common-utils";
-import {
 	DataProcessingError,
 	extractSafePropertiesFromMessage,
-} from "@fluidframework/container-utils";
+	ITelemetryLoggerExt,
+} from "@fluidframework/telemetry-utils";
+import { assert, performance } from "@fluidframework/common-utils";
 import { isRuntimeMessage } from "@fluidframework/driver-utils";
 import { DeltaScheduler } from "./deltaScheduler";
 import { pkgVersion } from "./packageVersion";
