+++ packages/runtime/container-runtime/src/test/containerRuntime.spec.ts
@@ -19,12 +19,12 @@ import {
 } from "@fluidframework/runtime-definitions";
 import {
 	ConfigTypes,
+	DataProcessingError,
 	GenericError,
 	IConfigProviderBase,
 	mixinMonitoringContext,
 	MockLogger,
 } from "@fluidframework/telemetry-utils";
-import { DataProcessingError } from "@fluidframework/container-utils";
 import { MockDeltaManager, MockQuorumClients } from "@fluidframework/test-runtime-utils";
 import { IContainerRuntime } from "@fluidframework/container-runtime-definitions";
 import { IErrorBase, IRequest, IResponse, FluidObject } from "@fluidframework/core-interfaces";
