+++ packages/test/test-end-to-end-tests/src/test/batchBreakRegression.spec.ts
@@ -12,14 +12,13 @@ import {
 import { requestFluidObject } from "@fluidframework/runtime-utils";
 import { ITestObjectProvider, TestFluidObject, timeoutPromise } from "@fluidframework/test-utils";
 import { describeNoCompat, itExpects } from "@fluid-internal/test-version-utils";
-import { isILoggingError } from "@fluidframework/telemetry-utils";
+import { DataProcessingError, isILoggingError } from "@fluidframework/telemetry-utils";
 import { TypedEventEmitter } from "@fluidframework/common-utils";
 import {
 	IDocumentMessage,
 	ISequencedDocumentMessage,
 	ISequencedDocumentSystemMessage,
 } from "@fluidframework/protocol-definitions";
-import { DataProcessingError } from "@fluidframework/container-utils";
 import { IContainerRuntimeOptions } from "@fluidframework/container-runtime";
 
 /**
