+++ packages/utils/telemetry-utils/package.json
@@ -69,6 +69,7 @@
 		"@fluidframework/common-utils": "^1.1.1",
 		"@fluidframework/core-interfaces": "workspace:~",
 		"@fluidframework/core-utils": "workspace:~",
+		"@fluidframework/protocol-definitions": "^1.1.0",
 		"debug": "^4.1.1",
 		"events": "^3.1.0",
 		"uuid": "^9.0.0"
